Reorganize files and imports
