+++ agency-dashboard/package.json
@@ -51,7 +51,7 @@
     "customize-cra": "^1.0.0",
     "jest-fetch-mock": "^3.0.3",
     "react-app-rewired": "^2.2.1",
-    "react-scripts": "5.0.1",
+    "react-scripts": "^5.0.1",
     "resize-observer-polyfill": "^1.5.1"
   },
   "config-overrides-path": "../config-overrides.js",
