+++ agency-dashboard/src/DashboardView/DashboardView.test.tsx
@@ -15,11 +15,10 @@
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 // =============================================================================
 import { render, screen } from "@testing-library/react";
-import React from "react";
 import { MemoryRouter } from "react-router-dom";
 
-import DashboardView from "./DashboardView";
-import { StoreProvider } from "./stores";
+import { StoreProvider } from "../stores";
+import { DashboardView } from ".";
 
 beforeEach(() => {
   fetchMock.resetMocks();
