+++ agency-dashboard/src/index.tsx
@@ -36,6 +36,7 @@ function ScrollToTop() {
 const root = ReactDOM.createRoot(
   document.getElementById("root") as HTMLElement
 );
+
 root.render(
   <React.StrictMode>
     <GlobalStyle />
